pyproject.toml
[project]
name = "kis-github"
version = "1.0.0"
description = "korea investment open api sample codes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.1",
    "pycryptodome>=3.23.0",
    "pyqt6>=6.9.1",
    "pyside6>=6.9.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "websockets>=15.0.1",
    "python-telegram-bot>=20.0.0",
]
